Rename segmentation to semantic_segmentation; Improve CodeStyle;
