+++ examples/experimental/onnx/semantic_segmentation/ac_configs/icnet_camvid.yml
