+++ examples/experimental/onnx/semantic_segmentation/onnx_ptq_segmentation.py
@@ -24,7 +24,7 @@ from nncf.experimental.onnx.dataloaders.segmentation_dataloader import create_da
 
 
 def run(onnx_model_path: str, output_model_path: str, dataset_name: str,
-        dataset_path: str, batch_size: int, shuffle: bool, num_init_samples: int,
+        dataset_path: str, num_init_samples: int,
         input_shape: List[int], ignored_scopes: List[str] = None):
     print("Post-Training Quantization Parameters:")
     print("  number of samples: ", num_init_samples)
@@ -34,8 +34,7 @@ def run(onnx_model_path: str, output_model_path: str, dataset_name: str,
     print(f"The model is loaded from {onnx_model_path}")
 
     # Step 1: Initialize the data loader.
-    dataloader = create_dataloader_from_segmentation_torch_dataset(dataset_name, dataset_path, input_shape,
-                                                                   batch_size=batch_size, shuffle=shuffle)
+    dataloader = create_dataloader_from_segmentation_torch_dataset(dataset_name, dataset_path, input_shape)
 
     # Step 2: Create a pipeline of compression algorithms.
     builder = CompressionBuilder()
@@ -69,8 +68,6 @@ if __name__ == '__main__':
     parser.add_argument("--data",
                         help="Path to dataset",
                         type=str)
-    parser.add_argument("--batch_size", help="Batch size for initialization", default=1)
-    parser.add_argument("--shuffle", help="Whether to shuffle dataset for initialization", default=True)
     parser.add_argument("--input_shape", help="Model's input shape", nargs="+", type=int, default=[1, 3, 768, 960])
     parser.add_argument("--init_samples", help="Number of initialization samples", type=int, default=300)
     parser.add_argument("--ignored_scopes", help="Ignored operations ot quantize", nargs="+", default=None)
@@ -79,8 +76,6 @@ if __name__ == '__main__':
         args.output_model_path,
         args.dataset_name,
         args.data,
-        args.batch_size,
-        args.shuffle,
         args.init_samples,
         args.input_shape,
         args.ignored_scopes
