+++ tests/onnx/sanity_tests/test_classification_sanity_sample.py
@@ -65,7 +65,7 @@ def mock_dataloader_creator(dataset_path, input_shape, batch_size, shuffle):
 
 @pytest.mark.parametrize(("model, input_shape"),
                          zip(MODELS, INPUT_SHAPES))
-@patch('examples.experimental.onnx.onnx_ptq_classification.create_dataloader_from_imagenet_torch_dataset',
+@patch('examples.experimental.onnx.classification.onnx_ptq_classification.create_dataloader_from_imagenet_torch_dataset',
        new=mock_dataloader_creator)
 def test_sanity_quantize_sample(tmp_path, model, input_shape):
     model_name = str(model.__class__)
